WFP-1126 get institutional report count from staging database (#126)
* WFP-1126 get institutional report count from staging database

* WFP-1126 use parole reports from WMT staging in workload calculation

* WFP-1126 rename from institutionalReport to paroleReport

* WFP-1126 do not default pareleReports to 0

* WFP-1126 test when no institutional report at all

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsworkload/jpa/entity/WorkloadPointsEntity.kt
@@ -61,7 +61,7 @@ data class WorkloadPointsEntity(
   val fastCourtReportPoints: BigInteger,
 
   @Column(name = "parom")
-  val paroleReportWeighting: BigInteger,
+  val paroleReportWeighting: Int,
 
   @Column(name = "paroms_enabled")
   val paroleReportWeightingEnabled: Boolean,
